MCLOUDMCLOUD-5352: [ece-tools] Create a functional test for Split DB functionality
